refactor(compact): move operation-log/compact to op-log/persistence/compact
