+++ src/app/op-log/persistence/compact/action-type-codes.spec.ts
