+++ src/app/op-log/sync-providers/file-based/file-based-sync.types.ts
@@ -3,8 +3,8 @@ import {
   type FileBasedSyncData as GenericFileBasedSyncData,
   type SyncFileCompactOp as GenericSyncFileCompactOp,
 } from '@sp/sync-providers/file-based';
-import type { CompactOperation } from '../../../core/persistence/operation-log/compact/compact-operation.types';
 import type { ArchiveModel } from '../../../features/time-tracking/time-tracking.model';
+import { CompactOperation } from '../../persistence/compact/compact-operation.types';
 
 export { FILE_BASED_SYNC_CONSTANTS };
